+++ src/paperang_cli/__init__.py
@@ -1,7 +1,7 @@
 """paperang-cli package."""
 
 from paperang_cli._version import __version__
-from paperang_cli.api import PaperangP1
+from paperang_cli.api import PaperangP1, PaperangP2
 from paperang_cli.config import PaperangCliConfig
 from paperang_cli.errors import ConfigError, DriverError, PaperangCliError, PrinterNotFoundError, SafetyError
 from paperang_cli.models import BatteryStatus, BluetoothMacStatus, PrintResult, PrinterDevice, PrinterStatus, ProbeResult
@@ -15,6 +15,7 @@ __all__ = [
 	"PaperangCliConfig",
 	"PaperangCliError",
 	"PaperangP1",
+	"PaperangP2",
 	"PrintResult",
 	"PrinterDevice",
 	"PrinterNotFoundError",